Address comments and split tests into files
