+++ src/NetAnalyzers/UnitTests/Microsoft.NetCore.Analyzers/Runtime/DetectPreviewFeatureTests.Methods.cs
@@ -1,0 +1,239 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System.Threading.Tasks;
+using Xunit;
+using VerifyCS = Test.Utilities.CSharpCodeFixVerifier<
+    Microsoft.NetCore.Analyzers.Runtime.DetectPreviewFeatureAnalyzer,
+    Microsoft.CodeAnalysis.Testing.EmptyCodeFixProvider>;
+
+namespace Microsoft.NetCore.Analyzers.Runtime.UnitTests
+{
+    public partial class DetectPreviewFeatureUnitTests
+    {
+        [Fact]
+        public async Task TestGenericPreviewParametersToPreviewMethod()
+        {
+            var csInput = @" 
+using System.Runtime.Versioning; using System;
+using System.Collections.Generic;
+namespace Preview_Feature_Scratch
+{
+
+    class Program
+    {
+        public Dictionary<int, Foo> {|#1:Getter|}(Dictionary<int, Foo> {|#0:foo|}) // Highlight Foo in the parameter list and return type here
+        {
+            return foo;
+        }
+
+        static void Main(string[] args)
+        {
+        }
+    }
+
+    [RequiresPreviewFeatures]
+    public class Foo
+    {
+    }
+}";
+
+            var test = TestCS(csInput);
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.UsesPreviewTypeParameterRule).WithLocation(0).WithArguments("Getter", "Foo"));
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.MethodReturnsPreviewTypeRule).WithLocation(1).WithArguments("Getter", "Foo"));
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestPreviewParametersToPreviewMethod()
+        {
+            var csInput = @" 
+using System.Runtime.Versioning; using System;
+namespace Preview_Feature_Scratch
+{
+
+    [RequiresPreviewFeatures]
+    class Program
+    {
+        public Foo Getter(Foo foo)
+        {
+            return foo;
+        }
+
+        static void Main(string[] args)
+        {
+            Program prog = new Program();
+            prog.Getter(new Foo());
+        }
+    }
+
+    [RequiresPreviewFeatures]
+    public class Foo
+    {
+    }
+}";
+
+            var test = TestCS(csInput);
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestPreviewParametersToMethods()
+        {
+            var csInput = @" 
+using System.Runtime.Versioning; using System;
+namespace Preview_Feature_Scratch
+{
+
+    class Program
+    {
+        public Foo {|#2:Getter|}(Foo {|#0:foo|})
+        {
+            return foo;
+        }
+
+        static void Main(string[] args)
+        {
+            Program prog = new Program();
+            prog.Getter({|#1:new Foo()|});
+        }
+    }
+
+    [RequiresPreviewFeatures]
+    public class Foo
+    {
+    }
+}";
+
+            var test = TestCS(csInput);
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.MethodUsesPreviewTypeAsParameterRule).WithLocation(0).WithArguments("Getter", "Foo"));
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.GeneralPreviewFeatureAttributeRule).WithLocation(1).WithArguments("Foo"));
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.MethodReturnsPreviewTypeRule).WithLocation(2).WithArguments("Getter", "Foo"));
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestUnmarkedPreviewMethodCallingPreviewMethod()
+        {
+            var csInput = @" 
+        using System.Runtime.Versioning; using System;
+        namespace Preview_Feature_Scratch
+        {
+        [RequiresPreviewFeatures]
+        public class Program
+        {
+            public bool CallSite()
+            {
+                return UnmarkedPreviewClass.SomeStaticMethod();
+            }
+        }
+
+        public class UnmarkedPreviewClass
+        {
+                [RequiresPreviewFeatures]
+                public static bool SomeStaticMethod()
+                {
+                    return false;
+                }
+        }
+        }
+        ";
+
+            var test = TestCS(csInput);
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestPreviewMethodCallingPreviewMethod()
+        {
+            var csInput = @" 
+        using System.Runtime.Versioning; using System;
+        namespace Preview_Feature_Scratch
+        {
+        public class Program
+        {
+            [RequiresPreviewFeatures]
+            public virtual void PreviewMethod()  { }
+
+            [RequiresPreviewFeatures]
+            void CallSite()
+            {
+                PreviewMethod();
+            }
+        }
+        }
+        ";
+
+            var test = TestCS(csInput);
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestMethodInvocation_Simple()
+        {
+            var csInput = @" 
+        using System.Runtime.Versioning; using System;
+        namespace Preview_Feature_Scratch
+        {
+
+            public class Program
+            {
+                [RequiresPreviewFeatures]
+                public virtual void PreviewMethod()
+                {
+
+                }
+
+                static void Main(string[] args)
+                {
+                    var prog = new Program();
+                    {|#0:prog.PreviewMethod()|};
+                }
+            }
+        }";
+
+            var test = TestCS(csInput);
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.GeneralPreviewFeatureAttributeRule).WithLocation(0).WithArguments("PreviewMethod"));
+            await test.RunAsync();
+        }
+
+        [Fact]
+        public async Task TestMethodInvocation_DeclareDerivedMethod()
+        {
+            var csInput = @" 
+        using System.Runtime.Versioning; using System;
+        namespace Preview_Feature_Scratch
+        {
+
+            public class Program
+            {
+                [RequiresPreviewFeatures]
+                public virtual void PreviewMethod()
+                {
+
+                }
+
+                static void Main(string[] args)
+                {
+                }
+            }
+
+            public class Derived : Program
+            {
+                public Derived() : base()
+                {
+                }
+
+                public override void {|#0:PreviewMethod|}()
+                {
+                    {|#1:base.PreviewMethod()|};
+                }
+            }
+        }";
+
+            var test = TestCS(csInput);
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.OverridesPreviewMethodRule).WithLocation(0).WithArguments("PreviewMethod", "Program.PreviewMethod"));
+            test.ExpectedDiagnostics.Add(VerifyCS.Diagnostic(DetectPreviewFeatureAnalyzer.GeneralPreviewFeatureAttributeRule).WithLocation(1).WithArguments("PreviewMethod"));
+            await test.RunAsync();
+        }
+    }
+}
